pyproject.toml
[tool.poetry]
name = "epifor"
version = "0.1.0"
description = "Tools to work with and model COVID 19 data"
authors = ["Tomáš Gavenčiak <gavento@gmail.com>"]
license = "MIT"
packages = [
    { include = "epifor" },
]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.1"
pyarrow = "^0.16.0"
requests = "^2.23.0"
numpy = "^1.18.1"
python-dateutil = "^2.8.1"
unidecode = "^1.1.1"
google-cloud-firestore = "^1.6"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
pylint = "^2.4.4"
pytest = "^5.3.5"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
